Remove deprecated rpc package and update api address
